Move things into correct folder
I had to move the ArmSubsystem and ReadMe.md file into a different
folder because there was two subsystem folders on github. Therefore
I had to remove the directory, but not delete the files. So, I moved
the two files into a different folder, deleted the one folder, and this
commit moves everything back into the correct folder.