Add a path follow state
Add a command for following a given path.
Either will set the odometry to where the robot is on the field, if it
can see an april tag, or it will set it to the start position of the
path. The purpose of setting it based on april tag is to help with
positioning of the robot, so the people placing it on the field do not
have to be as accurate at the beginning of the match. Of course, it is
still helpful, to place it in the right place and rotation, but it is
not as important as previous years.