Add QOL improvements (#1)
* Add new absolute encoder for future

Added a MAVCoder. This is a custom encoder created by another member on
the team, and creating an interface for it. Currently has no testing.

* Split constants into related files

Split all constants into a constants folder and related constants files
so that way if quick code debugging needs to happen in the pit, one
doesn't need to look through a massive constants file. Instead, find the
related file, and change the value.

* Remove Constants file

Removed Constants file, and changed all imports to use the new constants
file system.

* Move calls to related places and update library

Moved drivetrain set max speeds call to constructor of drivetrain
Split up the configurations into multiple different methods.
 - Configuring the controllers themselves (interfacing with the physical
   controller)
 - Configuring driver controller bindings
 - Created operator controller and method for configuring operator
   operator bindings
 - Created controllers for demo purposes, and methods for configuring
   their bindings
Updated external library to current version

* Add auto selection and telemetry constants file

Added a way for selecting which auto will be run.
 - Just a sendable chooser for commands

Added a file for all telemetry constants. This is where basically
everything that will be used for communicating between the robot and
computer (excluding controllers) will be housed.

* Fix issues that came up in testing