Add method for selecting paths based on alliance
Added a class based method for path selection.

The concept is that you have a class that loads all the paths for a
single alliance, you construct that class twice, once for each alliance,
and have a global reference to one of the two. At some point, you will
check what alliance you are on, and set the global reference to the
correct class.

The reason we did this is it is simple in concept, and execution, it
removes a lot of common code between the path loading, it is very
easily idiot proofed, and it ensures that if you have a path for one
alliance, you have a path for the other one as well.

The paths that were previously on this branch were removed, and paths
tailored to the system that we are currently using were added.