Fix auto runtime bugs
Fixed a bug with the theta controller. The trapezoid profile wasn't
allowing the controller to use the correct error initially, causing
rotation in one direction the first time auto was run.

Fixed a bug with checking whether the rotations are equal or not.
Originally, we were checking straight delta with the angle. This caused
issues sometimes where the angle would be very close to 2pi, and the
expected angle would be 0, therefore, this causes an error of 2pi, not
almost 0.