Get auto to a semi-reliable place (#9)
* Add method for selecting paths based on alliance

Added a class based method for path selection.

The concept is that you have a class that loads all the paths for a
single alliance, you construct that class twice, once for each alliance,
and have a global reference to one of the two. At some point, you will
check what alliance you are on, and set the global reference to the
correct class.

The reason we did this is it is simple in concept, and execution, it
removes a lot of common code between the path loading, it is very
easily idiot proofed, and it ensures that if you have a path for one
alliance, you have a path for the other one as well.

The paths that were previously on this branch were removed, and paths
tailored to the system that we are currently using were added.

* Change paths and loading

Changed the paths to have names that match what they are doing.
Added all the new paths to TrajectoryContainer
Added acceleration constants to DrivetrainConstants

* Update auto commands

Updated AprilRunCommand to use the new path names
Created a ChargeCircleCommand to use the new paths
Upadted RobotContainer to add a new auto option of ChargeCircleCommand

* Fix auto initialization bug

Fixed a bug where the program would crash if we enabled in auto twice
without restarting code in between enables
Moved resetting robot pose for auto to be in *one* location, rather than
many different locations

* Use PathPlanner for odometry updates during teleop

* Fix auto runtime bugs

Fixed a bug with the theta controller. The trapezoid profile wasn't
allowing the controller to use the correct error initially, causing
rotation in one direction the first time auto was run.

Fixed a bug with checking whether the rotations are equal or not.
Originally, we were checking straight delta with the angle. This caused
issues sometimes where the angle would be very close to 2pi, and the
expected angle would be 0, therefore, this causes an error of 2pi, not
almost 0.