Add rotation lock state (#10)
* Add RotationLockState

Added a rotation lock state for times when we want the driver to be
able to drive around, but not rotate, staying fixed at a certain angle.
I decided against just having zero rotational velocity, because that
doesn't guarantee no rotation. The modules can be a tiny bit cockeyed,
and over time, that will rotate the robot. Therefore, I am relying on
odometry being accurate, and correcting based on that.

* Debug rotation lock state