General Fix (#18)
* Add method for selecting paths based on alliance

Added a class based method for path selection.

The concept is that you have a class that loads all the paths for a
single alliance, you construct that class twice, once for each alliance,
and have a global reference to one of the two. At some point, you will
check what alliance you are on, and set the global reference to the
correct class.

The reason we did this is it is simple in concept, and execution, it
removes a lot of common code between the path loading, it is very
easily idiot proofed, and it ensures that if you have a path for one
alliance, you have a path for the other one as well.

The paths that were previously on this branch were removed, and paths
tailored to the system that we are currently using were added.

* Change paths and loading

Changed the paths to have names that match what they are doing.
Added all the new paths to TrajectoryContainer
Added acceleration constants to DrivetrainConstants

* Update auto commands

Updated AprilRunCommand to use the new path names
Created a ChargeCircleCommand to use the new paths
Upadted RobotContainer to add a new auto option of ChargeCircleCommand

* Fix auto initialization bug

Fixed a bug where the program would crash if we enabled in auto twice
without restarting code in between enables
Moved resetting robot pose for auto to be in *one* location, rather than
many different locations

* Use PathPlanner for odometry updates during teleop

* Fix auto runtime bugs

Fixed a bug with the theta controller. The trapezoid profile wasn't
allowing the controller to use the correct error initially, causing
rotation in one direction the first time auto was run.

Fixed a bug with checking whether the rotations are equal or not.
Originally, we were checking straight delta with the angle. This caused
issues sometimes where the angle would be very close to 2pi, and the
expected angle would be 0, therefore, this causes an error of 2pi, not
almost 0.

* Speed the robot up during auto

* Merge branch 'main' into Auto

* Change paths

Added paths for a one cone auto and removed paths that were used for
testing auto and odometry

* Update size of robot for practice bot

* Update trajectory container with new path

* Remove auto testing commands

* Add command for a one cone auto

* Add one cone command to auto chooser

* Update WPILib

* Add boolean for if the practice bot is being used

* Use practice bot boolean for drivetrain offsets

Used the practice bot boolean to check if we are currently using the
practice robot, and if so, use one set of offsets, otherwise, use a
different set.
Also changed drivetrain to use those offsets

* Updated limelights

Updated firmware on limelights, so the Limelight helper class was
updated to match a change
Added the other two limelights we are placing on the robot
Renamed THREE to CENTER to reflect where the limelight is placed

* Change ArmSubsystem to handle safety

Changed ArmSubsystem to automatically handle safety checking and running
 - Elevator will not move unless the arm is in a safe position
 - If the arm is moving to an unsafe position, move the elevator first
 - Otherwise, move the elevator and arm concurrently
Converted ArmSetpointCommand to a state
Simplified ArmSetpointState

* Make one cone paths accurate to alliance

* Tweak OneConeCommand

Made the arm actively hold onto the cone when moving to place it
Made the robot wait less time to place the cone

* Add paths for moving to a second cone

* Add a command for two cones

Made a command for placing the first cone, and moving over to the
second cone

* Add two cone command to shuffleboard

* Merge branch 'main' into Auto

* Fix merge with main

* Rename claw to intake in all places

* Remove unused paths

* Add path for picking up second cone

* Retune rotary arm pid and fix setpoints

Retuned rotary arm pid using Sysid
Fixed LOW_CONE and HUMAN_PLAYER_CONE setpoints
Added a ZERO setpoint for quick resets

* Make arm subsystem deal with arm safety

Changed the way that arm subsystem works so that the setting of arm and
elevator automatically handles the necessary safety for not tearing the
robot apart

* Update external library

Updated the library and added generics to use it

* Change cube current limits

Changed the cube current limits because we fixed one facet of the
mechanical issue that was causing us to shred cubes

* Change setpoints

* Add printouts for arm testing

* Update external library

* Fix bad merge

* Change cube grab current limits

* Remove unused field

* Switch to using comp bot

* Move and add constants

Moved nominal voltage to RobotConstants because that is where it felt
like it should go
Added a default current limit constant
Going with the idea that everything non-programmers should ever have to
change is in the constants folder

* Add intake constants

Added constants for cube and cone current limits, and used them in code
Removed shuffle board tables import because it was not being used

* Add ability to get true reading

Added the ability to get a reading from the MAVCoder that is not offset
for easier figuring out of what offsets for drivetrain should be

* Update drivetrain offsets

* Update ground cube setpoint

* Overhaul motor builder

Fixed nominal voltage input
Imported default current limit
Made the spark encoder and pid controller fields on the class

* Add ability to correct swerve modules

Added ability for the swerve module to know whether or not it is where
it should be, and if it isn't, correct itself.

* Update to current version of library

* Have swerve modules self-heal

Made the swerve modules check if the MAVCoder and NEO encoder are off
by more than a certain amount, and if so, reset the module