Add homing ability to arm
Added ability for the arm to home the elevator
The reason for this is so that way if something stupid happens, then
the elevator will eventually fall down because of gravity, the limit
switch will be tripped, and the elevator re-zeroed.
This will only happen once so we aren't messing with control loop things
while they are still running.