Change candle to be the singleton
Fixed construction of auto command, and fixed signaling state
Because the candle is no longer a subsystem, it will not cause issues
with causing unwanted canceling.
We will just need to be more careful with using the candle.
Technically, its just lights, so even if we have conflicting commands,
nothing on the robot itself will break, just have odd signaling.