Suppress unused method warning
This method will never actually be used in any commits that are pushed.
All of the uses will be outside of other commits, and cleaned up
afterwards.
This method is expected to be unused during commits, so this is just to
get rid of the yellow squiggle in the editor