Create abstract drive states
Created an abstract drive state.
Purpose for this is so that it is easier to do things with the
drivestate, without needing to do weird things with lambdas
Created a positional drive state to allow for controlling the position
of the robot in a state.
If wanted, you could only use the positional for certain components
and reoverride get speeds for the other components